requirements.txt
absl-py==0.12.0
anyio==2.2.0
argon2-cffi==20.1.0
astroid==2.5.1
astunparse==1.6.3
async-generator==1.10
asyncio==3.4.3
attrs==20.3.0
autopep8==1.5.5
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
blessed==1.18.0
brotlipy==0.7.0
cachetools==4.2.1
certifi==2020.12.5
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.0
enlighten==1.8.0
entrypoints==0.3
flake8==3.8.4
flatbuffers==1.12
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.32.0
h5py==2.10.0
iniconfig==1.1.1
ipykernel==5.5.0
ipympl==0.6.3
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
jedi==0.17.2
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==3.0.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.2
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
neovim==0.3.1
nest-asyncio==1.5.1
notebook==6.2.0
nptyping==1.4.0
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
outcome==1.1.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.6
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pydocstyle==5.1.1
pyflakes==2.2.0
pyfuse3==3.2.0
Pygments==2.8.1
pylint==2.7.2
pynvim==0.4.3
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
pytest-trio==0.7.0
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
requests-oauthlib==1.3.0
rope==0.18.0
rsa==4.7.2
ruamel-yaml==0.15.87
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
Send2Trash==1.5.0
sklearn==0.0
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcontainers==2.3.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
trio==0.18.0
typing-extensions==3.7.4.3
typish==1.9.1
ujson==4.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yapf==0.30.0
